backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
neomodel==5.2.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-django==4.7.0
gunicorn==21.2.0
watchdog==3.0.0
